Correct use of terms 'option' and 'argument'
Rename some functions and variables to correct the use of the terms
'argument', 'option' and 'parameter' according to the following
definitions:

A command line is split into an array of tokens called 'arguments'. An
'option' is a documented type of argument modifying the behavior of a
command. A 'parameter' is an argument that provides information to
either the command or one of its options.