Fix escaping of '.' in regular expressions
If a literal '.' is required in a regular expression, it must be written
either as "\\." (within double quotes) or as '\.' (within single
quotes). The latter is more readable, therefore change the quoting from
double to single quotes.

In addition, a Mac OS X version number should always start with '10'.
Therefore, the regular expression should reflect this.