Refactor check for supported installer applications
The verification of whether the specified installer application is
supported is based on the version of the Mac OS X included in the
installer application. This requires mounting the BaseSystem.dmg image
which is expected at a defined path within the mounted InstallESD.dmg
image. However, starting with macOS High Sierra (Mac OS X 10.13) the
path to the BaseSystem.dmg has changed. Therefore, the script fails
without warning before it can even check support for the installer
application.

To make the script more robust, it should not rely on assumptions about
the internal structure of the installer application when checking for a
supported installer application. This is done by checking the
'CFBundleDisplayName' of the installer application against a list of
supported display names of installer applications.