Add function to get installer application type
Starting with macOS High Sierra (Mac OS X 10.13) a new internal
structure of the installer application is introduced. To distinguish
this new type of installer application from the older type that is used
in the installer applications from Mac OS X Lion (Mac OS X 10.7) to
macOS Sierra (Mac OS X 10.12), a function is added which determines the
type of installer application.