Fix size calculation for writable disk image
The calculation for the required size of the writable disk image does
not take into account that the Packages directory and the BaseSystem.dmg
file must be copied to the writable disk image. Therefore, the
calculation must not subtract the size of the BaseSystem.dmg file from
the total number of non empty bytes of the InstallESD.dmg image.

Also updated the comments to better explain the calculation.