Fix output order in functions using output parsing
Since the function that performs the output parsing runs asynchronously
in a subshell, this function may still output text if the shell script
has already reached the next command in the script that also writes to
the output. This leads to garbage in the output.

To avoid this, we have to wait until the parsing function is finished.