Use a copy of the installer application if required
When using an Apple silicon Mac to create an ISO image for macOS 10.12
to 10.15, the required 'createinstallmedia' fails due to missing code
signatures for 'arm64'. As a workaround, a copy of the installer
application is created and the existing code signatures are replaced by
an ad-doc signature in this copy of the installer application.

The fix for the broken macOS Sierra installer is now also applied to a
copy of the installer application instead of temporarily modifying the
originally specified installer application.